fix: prioritize parental relationships from valid flowcells
there are duplicate entries for some subjects across flowcell identifiers, and the code is not smart
enough to determine from that which analyte should actually be flagged as a parent. improve parent
selection by filtering on flowcell IDs that have been requested at all in the set of target
subjects.