feat: split out joint calls into single sample vcfs with additional cleaning
this current implementation, as of the gatk LeftAlignAndTrimVariants step, emits vcfs that are not
compatible with Moon's limited vcf support. unfortunately, Moon doesn't bother to elaborate on what
aspect of the file it doesn't understand, and as such the options are either remove the gatk step
and revert to bcftools/glnexus only, or read through the Moon ruby source code and guess what the
issue is. I've done the ruby thing before, and it's a drag and very imprecise; so instead, I'm going
to remove the gatk step and switch to splitting out single-sample vcfs from joint calls of each
pedigree by itself.