fix: adjust region specification to make_examples, params completion for call_variants
an uncharacterized error arises when trying to call a subset of chrY with command line syntax
`--regions "chrY:start-stop"` syntax. the documentation suggests that this syntax should be
functional, but it seems to choke on it, and there are enough trivial differences between the
examples and the exact syntax that there's probably some absurd reason it doesn't like that exact
permutation of format. docs suggest it alternatively accepts bedfiles directly, so attempt to use
those instead of space-delimited CLI values. separately, call_variants wasn't getting an appropriate
params completion due to some esoteric behavior of snakemake's involving direct use of wildcard
completion in lambda/expand combinations. as such, I've fallen back to manually setting values in
the expand function based on the wildcards lambda parameter.