fix: add temporary improvement of project ID resolution logic for new style runs
the linker is causing real issues at this point, with its inability to access the expected project
ID for runs that originate outside of the legacy samplesheet. this patch does improve the logic
somewhat to prevent the worst collisions, but it still requires a longer term fix in handling
injecting new form project IDs into the linker, somehow.