feat: generalize read input to support either crams or bams
tools are emitting different output depending on whether they're provided bams or crams, which is
obviously fairly troubling. supporting bams or crams arbitrarily could cause issues, but for the
moment the only tool that rejects crams entirely is telomerecat, which is broken regardless.
eventually, there can be a utility that intercepts crams and converts them to temporary bams for
such tools, though that step would itself be extremely costly